+++ packages/terra-framework-docs/src/terra-dev-site/test/layout/LayoutDefault.test.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Layout from '../../../Layout';
-import TestHeader from '../common/TestHeader';
-import TestMenu from '../common/TestMenu';
-import TestContent from '../common/TestContent';
+import Layout from 'terra-layout';
+import TestHeader from './common/TestHeader';
+import TestMenu from './common/TestMenu';
+import TestContent from './common/TestContent';
 import styles from './TestLayoutCommon.module.scss';
 
 const cx = classNames.bind(styles);