+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 1.7.0 - (March 8, 2022)
 
 * Added
+  * Migrate docs from terra-layout
   * Added docs for `terra-filter-pills`
   * Added tests for `terra-filter-pills`